Add files via upload
This is a Hangman Game made using Python. In this game, the user has to guess the word from a bunch of words selected by the program. And if the user guesses the correct word, he/she wins and saves the man else the man dies. The words which are used in this program are present in a JSON file attached with the program.